moved polymer endpoint to existing novel_compound blueprints
