+++ Webapp/sources/blueprints/novel_compound/routes.py
@@ -53,3 +53,43 @@ def novel_compound(workgroup, workbook) -> Response:
             services.solvent.add(new_compound.name, new_compound.hazard_codes, nc)
 
     return jsonify({"feedback": new_compound.feedback})
+
+
+@novel_compound_bp.route("/_polymer_novel_compound", methods=["GET", "POST"])
+@login_required
+@workbook_member_required
+def polymer_novel_compound(workgroup, workbook) -> Response:
+    """
+    Adds a novel compound to the database and links it to the workbook of the current reaction.
+    The novel compound data is saved from the data the user enters in the form.
+    This function is triggered when a novel structure is drawn in the sketcher.
+
+    Args:
+        workgroup: the name of the workgroup that the workbook belongs to
+        workbook: the name of the workbook that we are looking for
+
+    Returns:
+        Flask Response with feedback about the operation.
+    """
+    # get the active workbook
+    workbook = services.workbook.get_workbook_from_group_book_name_combination(
+        workgroup, workbook
+    )
+
+    # check all the data the user has supplied to ensure it is valid and no unique constraints are broken.
+    new_compound = services.polymer_novel_compound.NewNovelCompound(workbook)
+    validation_results = new_compound.validate()
+    if validation_results == "validation_successful":
+        # create novel compound
+        services.polymer_novel_compound.add(
+            new_compound.name,
+            new_compound.mol_formula,
+            new_compound.mol_weight,
+            new_compound.density,
+            new_compound.concentration,
+            new_compound.hazard_codes,
+            new_compound.smiles,
+            new_compound.workbook.id,
+        )
+
+    return jsonify({"feedback": new_compound.feedback})
